+++ packages/schedule/package.json
@@ -24,6 +24,7 @@
   "exports": {
     ".": "./dist/index.js",
     "./nitro": "./dist/nitro.js",
+    "./runtime": "./dist/runtime.js",
     "./runtime/execute": "./dist/runtime/execute.js",
     "./vite": "./dist/vite.js",
     "./package.json": "./package.json"
