+++ packages/schedule/src/index.ts
@@ -1,20 +1,7 @@
 export { defineSchedule } from "./definition.ts"
 export type {} from "./registry-module.d.ts"
 export { ScheduleError } from "./errors.ts"
-export { schedules, validateRuntimeScheduleCron } from "./runtime/client.ts"
 export { createScheduleRun, executeStaticSchedule } from "./runtime/execute.ts"
-export { createMemoryRuntimeScheduleStore } from "./runtime/store.ts"
-export {
-  enterScheduleRuntimeEvent,
-  getRuntimeScheduleStore,
-  getScheduleRuntimeEvent,
-  getScheduleRuntimeRegistry,
-  loadScheduleDefinition,
-  resetScheduleRuntime,
-  runWithScheduleRuntimeEvent,
-  setRuntimeScheduleStore,
-  setScheduleRuntimeRegistry,
-} from "./runtime/state.ts"
 
 export type {
   DiscoveredScheduleDefinition,
