+++ packages/schedule/tsdown.config.ts
@@ -14,6 +14,7 @@ export default defineConfig({
   entry: [
     "src/index.ts",
     "src/nitro.ts",
+    "src/runtime.ts",
     "src/runtime/execute.ts",
     "src/vite.ts",
   ],
